v1.1.1 (Sun 01 Apr 2024) - refactor (move `Console Play` to example)
